pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qcrypto"
version = "0.1.0"
description = "Quantum-safe cryptography utilities built on top of liboqs-python"
readme = "README.md"
requires-python = ">=3.8"
license = { text = "MIT" }

authors = [
    { name = "Victor Postea" }
]

dependencies = [
    "liboqs-python>=0.14.0",
]

[tool.setuptools.packages.find]
where = ["src"]